+++ next.config.ts
@@ -3,6 +3,9 @@ import type { NextConfig } from "next";
 
 const nextConfig: NextConfig = {
   output: "standalone",
+  compiler: {
+    styledComponents: true,
+  },
   images: {
     remotePatterns: [
       {